+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - added: Done button is a part of SimpleTextInput
+- added: Ability to add a custom token for any wallet directly from CreateWalletSelectCryptoScene
 - changed: Better send error message for insufficient gas
 - changed: Upgrade to React Native v0.73.6
 - changed: "Split" wallet menu options consolidated into one